requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-redis
psycopg2-binary
requests
django-prometheus
python-json-logger
opentelemetry-api 
opentelemetry-sdk 
opentelemetry-instrumentation-django 
opentelemetry-exporter-jaeger
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-redis
opentelemetry-instrumentation-logging
deprecated
pytest
celery

coverage

flake8
black

isort

autoflake

dotenv